refactor(address-validator): introduce Validator interface, type all currency params
Replace validator: any on Currency with a typed Validator interface, making the
dispatch contract enforceable at compile time. Every module registered in
CURRENCIES[] must now satisfy { isValidAddress, getAddressType? } with proper
Currency-typed signatures — TypeScript catches missing or mis-typed implementations.

Also replaces all remaining currency?: any annotations in the seven multi-network
validators (ada, bch, bitcoin, loki, monero, sys, tron) that read from the currency
object at runtime, and _currency?: any in the 18 single-network validators where
the parameter is accepted but unused.

Currency and Validator interfaces live in a new currency-types.ts to avoid
a circular dependency (currencies.ts imports all validator modules, validators
import Currency).
