+++ packages/address-validator/src/atom_validator.ts
@@ -1,4 +1,5 @@
 import { addressType } from './crypto/utils';
+import type { Currency } from './currency-types';
 
 const ALLOWED_CHARS = 'qpzry9x8gf2tvdw0s3jn54khce6mua7l';
 
@@ -6,7 +7,7 @@ const regexp = new RegExp('^(cosmos)1([' + ALLOWED_CHARS + ']+)$');
 
 export const isValidAddress = (address: string): boolean => regexp.exec(address) !== null;
 
-export const getAddressType = (address: string, _currency?: any, _networkType?: string) => {
+export const getAddressType = (address: string, _currency?: Currency, _networkType?: string) => {
     if (isValidAddress(address)) {
         return addressType.ADDRESS;
     }
