+++ packages/address-validator/src/eos_validator.ts
@@ -1,4 +1,5 @@
 import { addressType } from './crypto/utils';
+import type { Currency } from './currency-types';
 
 function isValidEOSAddress(address: string): boolean {
     const regex = /^[a-z0-9]+$/g;
@@ -11,7 +12,7 @@ function isValidEOSAddress(address: string): boolean {
 
 export const isValidAddress = (address: string): boolean => isValidEOSAddress(address);
 
-export const getAddressType = (address: string, _currency?: any, _networkType?: string) => {
+export const getAddressType = (address: string, _currency?: Currency, _networkType?: string) => {
     if (isValidAddress(address)) {
         return addressType.ADDRESS;
     }
