+++ packages/address-validator/src/nano_validator.ts
@@ -2,6 +2,7 @@ import baseX from 'base-x';
 
 import * as cryptoUtils from './crypto/utils';
 import { addressType } from './crypto/utils';
+import type { Currency } from './currency-types';
 
 const ALLOWED_CHARS = '13456789abcdefghijkmnopqrstuwxyz';
 
@@ -27,7 +28,7 @@ export const isValidAddress = (address: string): boolean => {
     return false;
 };
 
-export const getAddressType = (address: string, _currency?: any, _networkType?: string) => {
+export const getAddressType = (address: string, _currency?: Currency, _networkType?: string) => {
     if (isValidAddress(address)) {
         return addressType.ADDRESS;
     }
