+++ packages/address-validator/src/nxt_validator.ts
@@ -1,10 +1,11 @@
 import { addressType } from './crypto/utils';
+import type { Currency } from './currency-types';
 
 const nxtRegex = new RegExp('^NXT(-[A-Z0-9]{4}){3}-[A-Z0-9]{5}$');
 
 export const isValidAddress = (address: string): boolean => nxtRegex.test(address);
 
-export const getAddressType = (address: string, _currency?: any, _networkType?: string) => {
+export const getAddressType = (address: string, _currency?: Currency, _networkType?: string) => {
     if (isValidAddress(address)) {
         return addressType.ADDRESS;
     }
