+++ packages/address-validator/src/solana_validator.ts
@@ -1,5 +1,6 @@
 import * as base58 from './crypto/base58';
 import { addressType } from './crypto/utils';
+import type { Currency } from './currency-types';
 
 export const isValidAddress = (address: string): boolean => {
     try {
@@ -11,7 +12,7 @@ export const isValidAddress = (address: string): boolean => {
     }
 };
 
-export const getAddressType = (address: string, _currency?: any, _networkType?: string) => {
+export const getAddressType = (address: string, _currency?: Currency, _networkType?: string) => {
     if (isValidAddress(address)) {
         return addressType.ADDRESS;
     }
