+++ packages/address-validator/src/zil_validator.ts
@@ -1,6 +1,7 @@
 import { bech32 } from '@scure/base';
 
 import { addressType } from './crypto/utils';
+import type { Currency } from './currency-types';
 
 const ALLOWED_CHARS = 'qpzry9x8gf2tvdw0s3jn54khce6mua7l';
 
@@ -16,7 +17,7 @@ export const isValidAddress = (address: string): boolean => {
     return !!(decoded && decoded.words.length === 32);
 };
 
-export const getAddressType = (address: string, _currency?: any, _networkType?: string) => {
+export const getAddressType = (address: string, _currency?: Currency, _networkType?: string) => {
     if (isValidAddress(address)) {
         return addressType.ADDRESS;
     }
